refactor: move authorization-server modules into authorization_server/
Continue the role-based namespacing. Move the provider-side modules into
oauth2_provider/authorization_server/ and the shared oauthlib bridge into
oauth2_provider/core/, with backward-compatible import shims at the old paths:

- authorization_server/: bcp, dcr, cimd, forms, admin
- core/backends_oauthlib (was oauth2_backends)

Shims alias the old module name via sys.modules so every attribute (including
private, underscore-prefixed names imported by tests) resolves with identical
object identity. The old paths emit a DeprecationWarning, except
oauth2_provider.admin, whose shim is silent because Django's admin
autodiscovery imports it at startup.

The DEFAULTS dotted-string defaults (OAUTH2_BACKEND_CLASS, DCR/CIMD classes,
*_ADMIN_CLASS) are repointed to the canonical paths; user overrides that still
name the old paths keep working through the shims.

Verified: full non-e2e suite passes (996), no first-party deprecation warnings
at startup/autodiscover, import-compat identity holds for old paths, and
makemigrations --check reports no changes.
