+++ oauth2_provider/decorators.py
@@ -4,7 +4,7 @@ from django.core.exceptions import ImproperlyConfigured
 from django.http import HttpResponse, HttpResponseForbidden
 from oauthlib.oauth2 import Server
 
-from .oauth2_backends import OAuthLibCore
+from .core.backends_oauthlib import OAuthLibCore
 from .oauth2_validators import OAuth2Validator
 from .core.scopes import get_scopes_backend
 from .settings import oauth2_settings
