+++ oauth2_provider/http.py
@@ -1,13 +1,14 @@
 """Backward-compatible import shim.
 
-``oauth2_provider.http`` moved to ``oauth2_provider.core.http`` when the package was reorganized by
-OAuth2 role. Importing from this old path still works but is deprecated and will
-be removed in django-oauth-toolkit 4.0.
+``oauth2_provider.http`` moved to ``oauth2_provider.core.http`` when the package was reorganized
+by OAuth2 role. Importing from this old path still works but is deprecated and
+will be removed in django-oauth-toolkit 4.0.
 """
 
+import sys
 import warnings
 
-from oauth2_provider.core.http import *  # noqa: F401,F403
+from oauth2_provider.core import http as _moved
 
 warnings.warn(
     "oauth2_provider.http has moved to oauth2_provider.core.http. The old import path is "
@@ -15,3 +16,7 @@ warnings.warn(
     DeprecationWarning,
     stacklevel=2,
 )
+
+# Alias the old module name to the moved module so every attribute (including
+# private, underscore-prefixed names) resolves with identical object identity.
+sys.modules[__name__] = _moved
