+++ oauth2_provider/views/base.py
@@ -20,7 +20,7 @@ from oauthlib.oauth2.rfc8628 import errors as rfc8628_errors
 
 from ..compat import login_not_required
 from ..exceptions import OAuthToolkitError
-from ..forms import AllowForm
+from ..authorization_server.forms import AllowForm
 from ..http import OAuth2ResponseRedirect
 from ..models import get_access_token_model, get_application_model, get_device_grant_model
 from ..oauth2_validators import is_valid_resource_uri
